requirements.txt

Flask>=3.0,<4
gunicorn>=21.2,<23
python-dotenv>=1.0,<2
redis>=5.0,<6
requests>=2.31,<3
requests-oauthlib>=2.0,<3
